feat(backend): 调整切换演练视图 #14694
